feature: apply BN update for SWA model
